CarPark/requirements.txt
djangorestframework==3.12.1
pep8==1.7.1
pylint==2.6.0
model-bakery==1.2.0
django-test-plus==1.4.0
pytest-django==3.10.0
pytest-sugar==0.9.4
pytest-cov==2.10.1